requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain-core>=0.2.38
langchain-openai>=0.1.6
langgraph>=0.2.0
pydantic>=2.7

# Additional dependencies
requests>=2.31.0
pyyaml>=6.0.1
